services/core-service/requirements.txt
# Core Service Dependencies
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic[email]==2.5.2
werkzeug==3.0.1
gunicorn==21.2.0
redis==5.0.1

# HTTP Client
httpx==0.27.0
tenacity==8.2.3
